\name{alka}
\docType{data}
\alias{alka}
\title{Alka-Seltzer dissoloving times}
\description{
This data set contains the times in seconds that it takes Alka-Seltzer tablets to dissolve in water and 7UP at two different temperatures.
}
\format{
  A data frame with 8 observations on the following 4 variables.
  \describe{
    \item{liquid}{: liquid (7UP or water)}
    \item{temp}{: temperature (cool or warm)}
    \item{time}{: time to dissolve (in seconds)}
    \item{block}{: bloaking level for 2x2 factorial design}
  }
}
\keyword{datasets}
