\name{bread}
\alias{bread}
\title{Bread rising experiment}
\description{
The data set bread contains height measurements of 48 cupcakes.
A batch of Hodgson Mill Wholesome White Bread mix was divided into three parts
and mixed with 0.75, 1.0, and 1.25 teaspoons of yeast, respectively.
Each part was made into 8 different cupcakes and baked at 350 degrees.
After baking, the height of each cupcake was measured.
Then the experiment was repeated at 450 degrees.
}
\format{
  A data frame with 48 observations on the following 3 variables.
  \describe{
    \item{yeast}{: quantity of yeast (.75, 1 or 1.25 teaspoons)}
    \item{temp}{: baking temperature (350 or 450 degrees)}
    \item{height}{: cupcake height}
  }
}
\keyword{datasets}
