\name{bread2}
\alias{bread2}
\title{Bread rising experiment}
\description{
The data set \code{bread2} contains averaged measurements
from the full data set, \code{bread}.
The 8 cupcakes in each temp/yeast combination have been averaged.}
\format{
  A data frame with 48 observations on the following 3 variables.
  \describe{
    \item{yeast}{: quantity of yeast (.75, 1 or 1.25 teaspoons)}
    \item{temp}{: baking temperature (350 or 450 degrees)}
    \item{height}{: cupcake height}
  }
}
\keyword{datasets}
