\name{cancer}
\docType{data}
\alias{cancer}
\title{Cigarette smoking and mortality}
\description{
The data set \code{cancer} examines a relationship between lung cancer
and cigarette smoking. The data consist of , a standardized measure of
smoking amount (smoke) and the standardized
mortality ratio (SMR) for males in England and Wales in 1970-72 who
were working in 25 different broad groups of jobs such as textile
workers, miners, etc. 
}
\format{
  A data frame with 25 observations on the following 2 variables.
  \describe{
    \item{smoke}{: standardized measure of smoking amount}
    \item{SMR}{: standardized mortality ratio}
  }
}
\source{
A Handbook of Small Data Sets by Hand, et al. (1994, p.67).
}
\keyword{datasets}
