\name{climate}
\docType{data}
\alias{climate}
\title{
Climate and geographical data for 50 of the largest US cities.
}
\description{
This data frame contains information about 50 of the largest US cities,
including location, rainfall, temperature and elevation.
}
\format{
  A data frame with 50 observations on the following 7 variables.
  \describe{
    \item{lat}{: latitude}
    \item{jan}{: average minimum January temperature (degrees F)}
    \item{rain}{: average rainfall in inches}
    \item{city}{: city names}
    \item{jul}{: average maximum July temperature}
    \item{elev}{: elevation above sea level in KW}
    \item{lon}{: longitude}
  }
}
\references{
The Universal Almanac (1992), ed. John W. Wright, Andrews and McNeel, Kansas City. 
}
\keyword{datasets}
