\name{college}
\docType{data}
\alias{college}
\title{
Statistics on colleges in 15 states.
}
\description{
The data frame college contains statistics relating
to colleges from 15 states. This is a sample of
fifteen states and certain statistics taken from
the Chronicle of Higher Education (most data is for
1992). All entries are in thousands so that Arkansas
(first row) has a population of 2,399,000, a yearly
per capita income of \$15,400, 85,700 undergraduates
students, 7,000 graduate students, and average cost
of tuition and fees at public universities of \$1,540,
and is located in the south (s for south).
}
\format{
  A data frame with 15 observations on the following 7 variables (all data in thousands).
  \describe{
    \item{school:}{State in which school is located.}
    \item{pop:}{State population.}
    \item{inc:}{Yearly per capita income.}
    \item{undergrad:}{Total number of undergraduate students.}
    \item{graduate:}{Total number of graduate students.}
    \item{fees:}{Average cost of tuition and fees.}
    \item{loc:}{Area of the country (s for south, w for west, ne for northeast, mw for midwest).}
  }
}
\keyword{datasets}
