\name{earthq}
\alias{earthq}
\docType{data}
\title{Earthquake data}
\description{
The data set \code{earthq} includes the dominant frequency and magnitude of
148 earthquakes.
}
\format{
  A data frame with 148 observations on the following 5 variables.
  \describe{
    \item{site}{: location of the earthquake}
    \item{freq}{: dominant frequency}
    \item{mag}{: magnitude}
    \item{depth}{: depth}
    \item{dist}{: distance}
  }
}
\source{
Earthquake Engineering and Structural Dynamics, Vol. 23, p. 583-597, 1994
}
\keyword{datasets}
