\name{etruscan}

\docType{data}

\alias{etruscan}

\title{

Skull widths for ancient Etruscans and modern Italians}

\description{

The data set \code{etruscan} contains the maximum width for 84 skulls

of Etruscan males and 70 modern Italian males. This data was gathered

in attempt to determine if Etruscans were native Italians or immigrants

from another land.

}

\format{

  A data frame with 154 observations on the following 2 variables.

  \describe{

    \item{width}{: skull width (in mm)}

    \item{group}{: ancient or modern}

  }

}

\source{

Medical Biology and Etruscan Origins, p. 136

}

\keyword{datasets}

