\name{ftp}

\docType{data}

\alias{ftp}

\title{

File transfer times}

\description{

Several students studied the relationship between file size and

transfer times using ftp (File Transfer Protocol) to retrieve files

from two Internet locations. At each location three different

files were transferred 5 times and averaged (to reduce variability).

}

\format{

  A data frame with 6 observations on the following 3 variables.

  \describe{

    \item{size}{: file size (in bytes)}

    \item{time}{: transfer time (in seconds)}

    \item{loc}{: internet location (0 or 1)}

  }

}

\keyword{datasets}

