\name{ftptime}
\docType{data}
\alias{ftptime}
\title{Forty FTP times}
\description{
The data in \code{ftptime} are 40 ftp times for a file of 343285 bytes
which was repeatedly obtained from a site in California.
}
\format{
  A numeric vector with 40 values.
}
\keyword{datasets}
