\name{insulate}

\alias{insulate}

\title{Effects of house insulation}

\description{

The data set \code{insulate} is one person's record of weekly gas

consumption (gas) and outside temperature (temp), before (insulation=0) 

and after (insulation=1) insulating a house. The house thermostat was

set at 20 degrees Celsius during the 26 weeks before and 30 weeks after

insulating.

}

\format{

  A data frame with 56 observations on the following 3 variables.

  \describe{

    \item{insulation}{: before insulation (0) or after (1)}

    \item{temp}{: outside temperature (in degrees Celsius)}

    \item{gas}{: gas consumption (in 1000 cubic feet)}

  }

}

\source{

A Handbook of Small Data Sets

}

\keyword{datasets}

