\name{magnet}
\alias{magnet}
\title{
Magnetic force of an electomagnet as a function of voltage and number of wire turns.
}
\description{
The magnet dataset is from an experiment concerning
the magnetic force of an electomagnet as a function
of voltage and number of wire turns. The device
was a wire wrapped around a core and measured at a
variety of voltages. The statistical design here is
actually a randomized complete block design where
the three eletromagnets are the blocks, and the
three voltages are levels of the factor voltage.
}
\format{
  \describe{
    \item{volt:}{Voltage applied (1.5 or 3.0 volts).}
    \item{turns:}{The number of wire turns (100, 200, or 300, as factors).}
    \item{force:}{The magnetic force.}
  }
}
\keyword{datasets}
