\name{marathon}
\docType{data}
\alias{marathon}
\title{
Winning times from New York City marathon}
\description{
The data consists of the winning times (in minutes)
for men and women at the New York Marathon, 1978-1998,
along with the temperature in Fahrenheit.
}
\format{
  A data frame with 21 observations on the following 4 variables.
  \describe{
    \item{year}{: 4-digit year}
    \item{temp}{: temperature (in Fahrenheit)}
    \item{mtime}{: men's winning time (in seconds)}
    \item{wtime}{: women's winning time (in seconds)}
  }
}
\source{
"The Effects of Temperature on Marathon Runners' Performance"
by David E. Martin and John F. Buoncristiani,
Chance, Vol. 12, No. 4, 1999
}
\keyword{datasets}
