\name{metalcut}

\docType{data}

\alias{metalcut}

\title{Metal cutting performance}

\description{

The data set \code{metalcut} is an attempt to determine which cutting

method (vertical or horizontal) yields the quickest and smoothest cuts

for three types of metal stock (angle, flat, round). Students from the

Biological Engineering Department measured cutting times and quality

for the six combinations of method and stock.

}

\format{

  A data frame with 18 observations on the following 4 variables.

  \describe{

    \item{cut}{: cutting method (hcut for horizontal or vcut for vertical}

    \item{stock}{: type of metal stock (angle, flat or round)}

    \item{time}{: cutting time (in seconds)}

    \item{quality}{: smooth (0) or rough (1)}

  }

}

\keyword{datasets}

