\name{monarch}

\alias{monarch}

\title{Post coronation lifespan}

\description{

The data set \code{monarch} contains the years lived after inauguration,

election, or coronation of popes, U.S. presidents, and British monarchs

from 1690 to 1970.

}

\format{

  A data frame with 72 observations on the following 3 variables.

  \describe{

    \item{group}{: group (K&Qs, popes or pres)}

    \item{years}{: year lived after coronation, inauguration or election}

    \item{name}{: name of the monarch, pope or president (no spaces)}

  }

}

\source{

\emph{Computer-Active Data Analysis} by Lunn and McNeil (1991)

 }

\keyword{datasets}

