\name{ncsu}
\alias{ncsu}
\title{
Number of degrees granted at ncsu by year.
}
\description{
The ncsu data frame has 92 rows and 2 columns. It gives
the number of degrees granted at North Carolina State
University (NCSU) from 1894 to 1983.
}
\format{
  A data frame with 92 observations on the following 2 variables.
  \describe{
    \item{year}{: 4 digit year}
    \item{degree}{: number of degrees granted}
  }
}
\keyword{datasets}
