\name{oil}
\alias{oil}
\docType{data}
\title{
Synthetic versus Conventional Oil
}
\description{
Synthetic versus Conventional Oil
}
\format{
  A data frame with 8 observations on the following 3 variables.
  \describe{
    \item{type}{: type of oil (conv or syn)}
    \item{vis}{: viscosity}
    \item{time}{: time}
  }
}
\keyword{datasets}
