\name{ozone}
\alias{ozone}
\title{
Monthly mean concentrations of ozone.
}
\description{
The ozone data frame has 552 rows and 3 columns. The
first column is the mean monthly ozone concentration
in Dobson units of the ozone layer at Arosa, Switzerland,
from 1926 to 1971.
}
\format{
  A data frame with 518 observations on the following 3 variables.
  \describe{
    \item{ozone}{: mean monthly ozone in Dobson units}
    \item{year}{: year in which measurements were taken (4 digits)}
    \item{month}{: month in which measurements were taken (3 letter abbreviation)}
  }
}
\keyword{datasets}
