\name{ping}

\docType{data}

\alias{ping}

\title{Internet ping times}

\description{

In Spring 2000 a team measured ping times of Internet servers at various

distances from Raleigh using a software program called NeoTrace. They

actually measured ping times at 4 different times of the day, but since

there was very little difference over time, we have averaged over the

times of day.

}

\format{

  A data frame with 12 observations on the following 2 variables.

  \describe{

    \item{dist}{: distance from Raleigh}

    \item{time}{: ping response time}

  }

}

\keyword{datasets}

