\name{popcorn}
\docType{data}
\alias{popcorn}
\title{
A 2x4 experiment to evaluate popcorn preparation.
}
\description{
The popcorn data frame has 16 rows and 6 columns. Four
factors are varied to see what produces the largest volume
of popcorn. The order that the observations were made is
in the column order.
}
\format{
  A data frame with 16 observations on the following 5 variables.
  \describe{
    \item{brand}{: brand of popcorn used (Orville Reddenbach or Jolly Time)}
    \item{temp}{: temperature of oil (hot or cold)}
    \item{quantity}{: quantity of oil used (3 Tsp or 1 Tsp)}
    \item{shake}{: whether the pan was shaken (yes or no)}
    \item{volume}{: volume of popcorn produced}
  }
}
\source{
This experiment was designed and carried out by Stan
Spencer for the NCSU statistics class ST516. Here is an
excerpt of his report:

INTRODUCTION

Popcorn has always been a crucial element of sustenance in
my life and I've always wondered what effects certain
factors have in the making of a good batch. Now, having
acquired some of the basic tools of statistical
experimentation I have been able to optimize my frequent
ritual of popping popcorn as well as understand exactly
how much of an effect these factors have on the desired
outcome. The purpose of the experiment was to optimize
the factors involved for the maximum volume of popcorn. I
focused on stovetop popping and didn't look at microwave
or air-pop methods.

DESIGN AND TEST CONDITIONS

I chose the two major popcorn brands with the motive of
trying to prove if Orville's claims are true. The second
factor refers to the temperature of the oil at the time
the popcorn was put in the pan. For the experiments
requiring cold oil I added the popcorn to the oil before
putting the pan on the stove. For the hot oil treatments
I let it heat for 20 seconds before adding the popcorn.
The quantity of oil factor required either one or three
tablespoons. The last factor I thought was important was
to either shake or not shake the pan during cooking. The
conditions of the test that were kept constant for each
treatment are: 1/2 cup of popcorn was used, the pan was
cooled and washed between treatments, the gas flame was
set to a constant, and the same pan and oil type were used
for each treatment. The volume was measured with a
measuring cup with units in mL.
}
\keyword{datasets}
