\name{precip}

\docType{data}

\alias{precip}

\title{

Average monthly rainfall

}

\description{

Average monthly rainfall in Raleigh, NC for 1948-1992.  Values are

recorded only for February, March, May, June, and August.

}

\format{

  A data frame with 45 observations on the following 6 variables.

  \describe{

    \item{year}{: 2 digit year}

    \item{feb}{: February rainfall (in inches)}

    \item{mar}{: March rainfall (in inches)}

    \item{jun}{: June rainfall (in inches)}

    \item{aug}{: August rainfall (in inches)}

    \item{may}{: May rainfall (in inches)}

  }

}

\keyword{datasets}

