\name{quake}

\docType{data}

\alias{quake}

\title{

Earthquake strengths and locations}

\description{

The data set \code{quake} contains the strengths of earthquakes measured

at the earth's continental plates. Much of the earth's seismic activity

is due to motion of the large plates that make up the crust of the earth.

Earthquakes occur when a buildup in tension between two layers of rock is

suddenly released. For this reason many earthquakes occur at plate

boundaries.

}

\format{

  A data frame with 496 observations on the following 5 variables.

  \describe{

    \item{lat}{: latitude of the event}

    \item{lon}{: longitude of the event}

    \item{direction}{: direction of earthquake}

    \item{strength}{: strength of the earthquake (Richter scale)}

    \item{plateid}{: numerical code for plate boundary}

  }

}

\keyword{datasets}

