\name{raleigh.snow}
\docType{data}
\alias{raleigh.snow}
\title{Annual snowfall in Raleigh, NC}
\description{
The data set \code{raleigh.snow} contains the annual snowfall totals
for Raleigh, NC from the 1962-63 season through the 1992-92 season.
}
\format{
  A data frame with 30 observations on the following 2 variables.
  \describe{
    \item{year}{: 2 digit year}
    \item{snow}{: annual snowfall (in inches)}
  }
}
\keyword{datasets}
