\name{raleigh.temp}
\docType{data}
\alias{raleigh.temp}
\title{
Average monthly temperatures, 1949-1988.
}
\description{
The raleigh.temp data frame has 480 rows and 3 columns.
Each year has 12 rows of data, one for each month. The
measurement is likely to be the average of the average
daily temperature, where the average daily temperature =
(daily high+daily low)/2.
}
\format{
  A data frame with 480 observations on the following 3 variables.
  \describe{
    \item{temp}{: temperature = (daily high+daily low)/2}
    \item{month}{: the month the measurements were taken during}
    \item{year}{: the year the measurements were taken during}
  }
}
\keyword{datasets}
