\name{Rlab internal}
\alias{Rlab.version}
\alias{cat.to.list}
\alias{describe}
\alias{describe2}
\alias{means.2way}
\alias{message2}
\alias{mfit.2way}
\alias{mklevel}
\alias{nplot}
\alias{random.walk}
\alias{read.data}
\alias{set.panel}
\alias{simulatereg}
\alias{simulatesamples}
\alias{simulatesums}
\alias{write.data}
\alias{ex}
\alias{airplane.ex}
\alias{bplot.ex}
\alias{capac.ex}
\alias{cavendish.ex}
\alias{climate.ex}
\alias{college.ex}
\alias{count.ex}
\alias{data.frame.ex}
\alias{gam.ex}
\alias{global.temp.ex}
\alias{lines.ex}
\alias{lm.ex}
\alias{lplot.ex}
\alias{magnet.ex}
\alias{means.ex}
\alias{michelson.ex}
\alias{mplot.ex}
\alias{ncsu.ex}
\alias{ozone.sw.ex}
\alias{plot.ex}
\alias{points.ex}
\alias{popcorn.ex}
\alias{predict.new.ex}
\alias{qunif.ex}
\alias{raleigh.temp.ex}
\alias{read.table.ex}
\alias{scan.ex}
\alias{slab.ex}
\alias{stats.ex}
\alias{title.ex}
\alias{unif.ex}
\alias{viscosity.ex}
\alias{weib.ex}
\alias{bplot.obj}
\alias{describe.bplot}
\alias{draw.bplot}
\alias{draw.bplot.obj}
\alias{stats.bplot}
\alias{dtri}
\alias{ptri}
\alias{qtri}
\title{
Rlab internal and secondary functions and datasets
}
\description{
Listed below are the internal and secondary functions and datasets in Rlab.

\code{Rlab.version}

\code{cat.to.list}
\code{describe}
\code{describe2}
\code{means.2way}
\code{message2}
\code{mfit.2way}
\code{mklevel}
\code{nplot}
\code{random.walk}
\code{read.data}
\code{set.panel}
\code{simulate.reg}
\code{simulate.samples}
\code{simulate.sums}
\code{write.data}

\code{ex}
\code{airplane.ex}
\code{bplot.ex}
\code{capac.ex}
\code{cavendish.ex}
\code{climate.ex}
\code{college.ex}
\code{count.ex}
\code{data.frame.ex}
\code{gam.ex}
\code{global.temp.ex}
\code{lines.ex}
\code{lm.ex}
\code{lplot.ex}
\code{magnet.ex}
\code{means.ex}
\code{michelson.ex}
\code{mplot.ex}
\code{ncsu.ex}
\code{ozone.sw.ex}
\code{plot.ex}
\code{points.ex}
\code{popcorn.ex}
\code{predict.new.ex}
\code{qunif.ex}
\code{raleigh.temp.ex}
\code{read.table.ex}
\code{scan.ex}
\code{slab.ex}
\code{stats.ex}
\code{title.ex}
\code{unif.ex}
\code{viscosity.ex}
\code{weib.ex}

\code{bplot.obj}
\code{bplot.old}
\code{describe.bplot}
\code{draw.bplot}
\code{draw.bplot.obj}
\code{stats.bplot}

\code{dtri}
\code{ptri}
\code{qtri}
}
\note{These are not intended for use by end users.}
\keyword{internal}
