\name{Rlab}
\alias{Rlab}
\title{
Rlab - Functions and Datasets for ST 370 at North Carolina State University 
}
\description{
Rlab is a collection of functions and datasets to be used
in the class ST 370, Probability and Statistics for Engineers,
at North Carolina State University.  For more information see
the class labs at:
\url{https://www4.stat.ncsu.edu/~bmasmith/NewST370WEB/rlab/rlab.html}

Some major methods include:
\itemize{
\item \code{bplot} - customized boxplot
\item \code{hplot} - customized boxplot
\item \code{lplot} - label plot (allows character z or y)
\item \code{means} - 2-way means
\item \code{mfit}  - 2-way interaction fit
\item \code{mplot} - means plot
\item \code{stats} - variety of statistics
\item \code{US}    - plot of the United States
\item \code{world} - plot of the world
}


These labs are based on Slab and Mlab by Doug Nychka and Dennis Boos.

DISCLAIMER:

This is software for statistical research and not for commercial uses. The
authors do not guarantee the correctness of any function or program in
this package. Any changes to the software should not be made without the
authors permission.
}
\keyword{datasets}
