\name{solar}
\docType{data}
\alias{solar}
\title{
Solar cell's output voltage
}
\description{
The data set \code{solar} is from an experiment investigating the effect
of surface area (cell), distance, and light intensity on the output voltage
of photovoltaic cells (solar cells).
}
\format{
  A data frame with 18 observations on the following 4 variables.
  \describe{
    \item{cell}{: surface area (0 for 8 sq. cm and 1 for 3 sq. cm)}
    \item{light}{: light intensity}
    \item{distance}{: distance}
    \item{volt}{: solar cell's output voltage}
  }
}
\keyword{datasets}
