\name{SpGravity}
\alias{SpGravity}
\title{
Specific gravity measurements
}
\description{
Data collected by a team of civil engineering students in an attempt to determine if the suggested 24 hour soak time for measuring the specific gravity and absorption of course aggregate (e.g., granite or limestone) was really necessary. They obtained samples of course aggregate from eight quarries and measured three types of specific gravity (BDSG, SSDSG and ASG), and absorption (abs) at five soak times (6, 18, 24, 48, and 72 hours). }
\format{
  A data frame with 40 observations on the following 7 variables.
  \describe{
    \item{quarry}{: quarry sample was taken from}
    \item{rock}{: type of rock (Granite or Limestone)}
    \item{time}{: soak time (6, 18, 24, 48, or 72 hours)}
    \item{BDSG}{: dry specific gravity}
    \item{SSDSG}{: SSD specific gravity}
    \item{ASG}{: apparent specific gravity}
    \item{abs}{: absorption}
  }
}
\keyword{datasets}
