\name{us.age}

\alias{us.age}

\title{American age data}

\description{

The data set \code{us.age} contains the average age for all Americans (all),

females (f), and males (m) for the years 1990-1999. Actually, the data from

the US Census Bureau are based on the 1990 census and then updated yearly.

}

\format{

  A data frame with 10 observations on the following 4 variables.

  \describe{

    \item{all}{: average age of all Americans}

    \item{f}{: average age of female Americans}

    \item{m}{: average age of male Americans}

    \item{year}{: 4 digit year}

  }

}

\source{

US Census Bureau

}

\keyword{datasets}

