\name{us.pop}

\docType{data}

\alias{us.pop}

\title{American population data}

\description{

The data set \code{us.pop} contains the United States population for the

years 1900-1999.

}

\format{

  A data frame with 100 observations on the following 2 variables.

  \describe{

    \item{year}{: 2 digit year}

    \item{pop}{: population (in millions)}

  }

}

\keyword{datasets}

