\name{viscosity}
\alias{viscosity}
\title{
Results from students experiments using water, oil and shampoo.
}
\description{
The viscosity data frame was the result of the
following ST370 experiment performed in the fall of
1996. The students' description is as follows.

For this experiment we used three different liquids:
water, cooking oil and shampoo. First we placed a
cup of shampoo in a microwave oven, and heated it
for 50 seconds. Immediately after that we transfered
the liquid to a dishwasher container. We turned
this container upside down with the spout closed and
poked a hole on the bottom part of it. Then we
placed a half cup measuring container beneath the
dishwasher container. Then we opened the spout of
the dishwashing container, and measured the time it
took for liquid to come out and fill the half cup
container. We repeated the same procedure with each
liquid three times. Then we placed the liquids at
room temperature in the container and repeated the
above prcedure three times as well. Then we placed
each liquid in the freezer 10 minutes at a time and
repeated the prior procedure three times.
}
\format{
  A data frame with 26 observations on the following 3 variables.
  \describe{
    \item{liquid}{: liquid used (shampoo, oil or water)}
    \item{temp}{: temperature (hot or cold)}
    \item{time}{: time (in seconds)}
  }
}
\keyword{datasets}
