\name{vocab}

\docType{data}

\alias{vocab}

\title{Children's vocabulary by age}

\description{

The data set \code{vocab} contains the average oral vocabulary size (words)

for children at different ages (age).

}

\format{

  A data frame with 10 observations on the following 2 variables.

  \describe{

    \item{age}{: age (in years)}

    \item{words}{: vocabulary (in number of words)}

  }

}

\source{

Discovering Psychology by Weiner, 1977

}

\keyword{datasets}

