\name{wire}

\docType{data}

\alias{wire}

\title{Wire resistance experiment}

\description{

The data set \code{wire} contains the results of an experiment to study

the relationship between the resistance of a wire and its gauge and its

length.

}

\format{

  A data frame with 27 observations on the following 3 variables.

  \describe{

    \item{gauge}{: wire gauge (in AWG units)}

    \item{length}{: wire length (in feet)}

    \item{resistance}{: resistance (in ohms)}

  }

}

\keyword{datasets}

