\name{xline}
\alias{xline}
\title{
  Draw a vertical line 
}
\description{
Adds  vertical lines in the plot region. 
}
\usage{
xline(x, ...)
}
\arguments{
\item{x}{
Values on x axis specifying location of vertical lines. 
}
\item{\dots}{
Any ploting options for abline. 
}
}
\seealso{
 yline, abline 
}
\examples{

plot( 1:10)
xline( 6.5, col=2)
 }
\keyword{aplot}
% docclass is function
% Converted by Sd2Rd version 1.21.
