\name{yarn}
\docType{data}
\alias{yarn}
\title{Yarn tensile strength experiment}
\description{
The data set \code{yarnred} is from a team project to investigate the effect
of count number and yarn type on the tensile strength of yarn. The Uster
Tensorapid 3, a standard machine for measuring tensile strength, was used
to take 10 measurements of strength from each of 6 yarn cones (one of each
type and count, 10 feet between measurements on each cone).
}
\format{
  A data frame with 60 observations on the following 3 variables.
  \describe{
    \item{count}{: thread count (24 or 36)}
    \item{type}{: yarn type (AirJet, OpenEnd or Ring)}
    \item{tensile}{: tensile strength}
  }
}
\keyword{datasets}
