\name{yline}
\alias{yline}
\title{
  Draw horizontal lines 
}
\description{
Adds  horizontal lines in the plot region. 
}
\usage{
yline(y, ...)
}
\arguments{
\item{y}{
Values on y axis specifying location of vertical lines. 
}
\item{\dots}{
Any ploting options for abline. 
}
}
\seealso{
 xline, abline 
}
\examples{
plot( 1:10)
yline( 4.0, col=3)
}
\keyword{aplot}
% docclass is function
% Converted by Sd2Rd version 1.21.
